requirements.txt
flask
opencv-python
mediapipe
numpy
pillow
tensorflow
imageio